% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door.playerchoice}
\alias{change_door.playerchoice}
\title{Change Door Function User Defined Scenario}
\usage{
change_door.playerchoice(stay = T, opened.doors, a.pick, doors)
}
\arguments{
\item{`stay`}{a true/false operator which determins if the user stays with the same initial selection,
or if a different door will be revealed for the final prize determination. Default setting is TRUE
for sticking with the initial option.
\code{opened.doors} passes all opened door information.
\code{a.pick} the initial pick the user made.
\code{doors} the vector determining how many doors there are in a given game.}
}
\value{
Returns the final door decision as a numeric double.
}
\description{
\code{change_door.playerchoice()} allows a user to opt into the stay strategy, or select
a previously unselected door.
}
\details{
This is the same as the base function, only using the user-defined scenario.
If a user switches doors, an unopened door is randomly selected from the remaining choices.
}
\examples{
new.game <- create_game.playerchoice(8, 5)
new.doors <- door_list.playerchoice(new.game)
first.pick <- select_door.playerchoice(new.doors)
goat.reveal <- open_goat_door.playerchoice(new.game, first.pick, new.doors)
final.choice <- change_door.playerchoice(stay=F, new.doors, first.pick, new.doors)
}
