% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Change selection to the other door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{`stay`}{a TRUE/FALSE condition indicating whether the contestant wants to
switch options. TRUE is the default condition.
\code{opened.door} passes a number through that was the revealed door.
\code{a.pick} is the number indicating the door the contestant initially selected.}
}
\value{
Returns a number between 1 and 3 as a numeric double type.
}
\description{
\code{change_door()} is a function that either keeps the initial pick or changes
the door selection.
}
\details{
#' This is the stage in the Monty Hall setup where the contestant is given the
opporunity to change their selection to the other unopened door.
}
\examples{
#' this.game <- create_game()
my.pick <- 3
open.sesame <- open_goat_door(this.game, my.pick)
change_door(stay=T, open.sesame, my.pick)
If the contestant decides to switch to the other door:
change_door(stay= F, open.sesame, my.pick)

}
