% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{create_game.playerchoice}
\alias{create_game.playerchoice}
\title{Create a user defined Monty Hall setup.}
\usage{
create_game.playerchoice(goats, cars)
}
\arguments{
\item{`goats`}{the number of goats in a game.
\code{cars} the number of cars in a game. Both parameters must be greater than 0.
Ideally there are a minimum of 2 goats to allow for the reveal logic to work for
the fully wrapped function.}
}
\value{
The function returns variable length character vector indicating the
positions of goats and the car. The length of the vector is determined by the number
of goats and cars.
}
\description{
\code{create_game.playerchoice()} allows users to create their own bespoke Monty Hall setup.
}
\details{
#' This function presents the user the opportunity to create a Monty Hall problem
with user defined numbers of goats and cars. The function also randomizes the order of
goats and cars u8sing the sample function.
}
\examples{
create_game.playerchoice(8,1)
create_game.playerchoice(4, 17)

}
