% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner.playerchoice}
\alias{determine_winner.playerchoice}
\title{Determine winner in player defined game.}
\usage{
determine_winner.playerchoice(final.pick, game)
}
\arguments{
\item{`final.pick`}{the prize door to be revealed.
\code{game} the game setup previously generated by the \code{create_gameplayerchoice()} function}
}
\value{
Returns a character type value.
}
\description{
Determines if the player has won or lost.
}
\details{
Returns the outcome of the user-definied Monty Hall Problem.
}
\examples{
new.game <- create_game.playerchoice(8, 5)
new.doors <- door_list.playerchoice(new.game)
first.pick <- select_door.playerchoice(new.doors)
goat.reveal <- open_goat_door.playerchoice(new.game, first.pick, new.doors)
final.choice <- change_door.playerchoice(stay=T, new.doors, first.pick, new.doors)
final.outcome <- determine_winner.playerchoice(final.choice, new.game)
}
