% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine win or loss of the game.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{`final.pick`}{is the final selection of the contestant as a number.
\code{game} is a vector that indicates the position of each prize.}
}
\value{
Returns a vector character.
}
\description{
Returns whether the game is won or lost.
}
\details{
Determines whether the player has successfully identified the door with the car,
or has revealed the other goat.
}
\examples{
test.game <- create_game()
first.pick <- select_door()
door.reveal <- open_goat_door(test.game, first.pick)
final.decision <- change_door(stay = T, door.reveal, first.pick)
outcome.stay <- determine_winner(final.decision, test.game)
outcome.stay

}
