% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{door_list.playerchoice}
\alias{door_list.playerchoice}
\title{Door list generator.}
\usage{
door_list.playerchoice(a.game)
}
\arguments{
\item{`a.game`}{is the user defined vector representing the}
}
\value{
Returns a vector with the number of doors.
}
\description{
Creates a vector of numbers based on user input.
}
\details{
Creates a vector with the appropriate number of doors as defined by the user-defined
scenario. This is largely a background function to make assigning and revealing
doors easier, especially in the fully nested function.
}
\examples{
new.game <- create_game.playerchoice(8, 5)
new.doors <- door_list.playerchoice(new.game)
}
