% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door.playerchoice}
\alias{open_goat_door.playerchoice}
\title{Open one goat door for the user defined game.}
\usage{
open_goat_door.playerchoice(game, a.pick, doors)
}
\arguments{
\item{`game`}{\itemize{
\item the vector defining the prize arrangement.
\code{a.pick} - the initial choice of door.
\code{doors} - the vector for the list of doors.
}}
}
\value{
The function returns a single number as a double type numeric between
one and \emph{n} doors from the \code{doors} parameter.
}
\description{
Returns an unopened goat door from a user-definied configuration as a numeric double.
}
\details{
Using the user-defined prize scenario, reveal a single goat door.
}
\examples{
new.game <- create_game.playerchoice(8, 5)
new.doors <- door_list.playerchoice(new.game)
first.pick <- select_door.playerchoice(new.doors)
goat.reveal <- open_goat_door.playerchoice(new.game, first.pick, new.doors)
}
