% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Reveal a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{The}{\code{game} parameter passes the configured game. \code{a.pick} passes the initial
door selection.}
}
\value{
The function returns a single number between 1 and 3. Also a double type numeric.
}
\description{
\code{open_goat_door()} reveals a door that the contestant has NOT selected and
contains a goat behind it.
}
\details{
In the secondary stage of the original Monty Hall problem, the host reveals
one of the remaining choices that is a losing door.
}
\examples{
this.game <- create_game()
my.pick <- 3
open_goat_door(this.game, my.pick)

}
