% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game.playerchoice}
\alias{play_game.playerchoice}
\title{Automated user-defined Monty Hall Game}
\usage{
play_game.playerchoice(goats, cars)
}
\arguments{
\item{`goats`}{the number of goats in a game.
\code{cars} the number of cars in a game. Both parameters must be greater than 0.
Ideally there are a minimum of 2 goats to allow for the reveal logic to work.
The function may misbehave otherwise if there are not at least 1 car and 2 goats, which is
equivalent to the original Monty Hall Problem setup.}
}
\value{
Returns a character list of both outcomes--stay or switch--in response to a user defined game.
}
\description{
This is the same as the base Monty Hall simulator function, but with user defined parameters
with regards to the number of goats and cars.
}
\details{
A fully nested and self-contained function that returns outcomes in a Monty Hall style problem.
The function returns the results for staying and changing doors.
}
\examples{
play_game.playerchoice(2, 3) \%>\% pander()  # pander is for pretty printing
}
