% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Nested Monty Hall Problem function.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function}
}
\value{
The function returns a list with a 2x2 data frame.
}
\description{
Automates the Monty Hall problem setup and returns the outcomes for both scenarios.
}
\details{
This function automates and returns the outcomes for staying with the initial decision
as well as the strategy to change doors. This allows the user to quickly iterate
and analyze numerous games without manually entering each individual game.
}
\examples{
outcome <- play_game()
outcome

}
