% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Iterates and stores simulation results for the Monty Hall problems.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{`n=100`}{n is the number of simulations to be run. The default value is 100.}
}
\value{
Returns the results as a data frame. Has dplyr dependencies.
}
\description{
\code{play_n_games()} allows users to loop the \code{play_game()} function for a defined
number of loops and then returns the summary of all outcomes.
}
\details{
Users can define a number of times to run the simulation. Each loop contains
the outcome for stay and changing door strategies and returns it into a list.
This allows for easy statistical analysis.
}
\examples{
play_n_games()
play_n_games( n = 5000)

}
