% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door.playerchoice}
\alias{select_door.playerchoice}
\title{Door selection for user-defined game.}
\usage{
select_door.playerchoice(door.order)
}
\arguments{
\item{`door.order`}{is a vector with the number of doors in the user-defined game scenario.}
}
\value{
Returns a number representing the initial door choice.
}
\description{
\code{select_door.playerchoice} randomly samples a number from the list of doors and returns one value.
}
\details{
Since in this version of the simulation, it is unknown what number of doors there will be,
the code allows for any number of doors to be presented for selection. Only one door
will be selected.
}
\examples{
new.game <- create_game.playerchoice(8, 5)
new.doors <- door_list.playerchoice(new.game)
first.pick <- select_door.playerchoice(new.doors)
}
