% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Randomly select a door for the contestant.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3 indicating the initial door selection,
specifically a double type numeric.
}
\description{
\code{select_door()} automatically and randomly samples a door from 1-3 as a character.
}
\details{
#' In the original Monty Hall premise the contestant is given the option to select a door.
This sets the original door for the game, and allows for comparison against the strategy
of picking a different door.
}
\examples{
select_door()

}
